Add role name tests and improve logging in RoleService
Enhanced the logging mechanism in RoleService.java for better traceability. In addition, several new unit tests were introduced in RoleServiceTest.java to validate the 'getRoleNamesForDbgapPermissions' method under various conditions. Transferred related tests from UserServiceTest.java to RoleServiceTest.java for better organization.
